Bug fix for insitution name not being populate on first login
This fixes a bug when using Auth0 where the user logs in the first time and they do not see their correct courses because their profile does not include the Institution name.

This bug fix manually sets the user Institution name on the first log in along with the other fields like first name, last name, and email. This is a bit of a hack because we are passing in that institution information through Auth0 which is not expected by Moodle. It expects people to be added to Institutions through other means apparently. This does fix the issue for us though.